requirements.txt
django==4.2.11
djangorestframework==3.15.1
gtts==2.5.1
django-cors-headers==4.3.1
transformers==4.38.2
torch==2.2.1+cu121
torchvision==0.17.1+cu121
torchaudio==0.2.2+cu121
soundfile==0.12.1
numpy==1.26.4
sentencepiece==0.2.0
python-dotenv==1.0.0